Revert "IGNITE-20764 Extract script parse into separate implementation."
This reverts commit 9119576d0e8cb13a2d51b7e01d70a056d7bc256d.
